+++ libs/components/datetime/src/lib/modules/datepicker/datepicker-calendar.component.ts
@@ -74,7 +74,7 @@ export class SkyDatepickerCalendarComponent {
 
   #config: SkyDatepickerConfigService;
 
-  public constructor(config: SkyDatepickerConfigService) {
+  constructor(config: SkyDatepickerConfigService) {
     this.#config = config;
     this.configureOptions();
   }