+++ libs/components/indicators/src/lib/modules/icon/types/icon-resolved.ts
@@ -1,0 +1,7 @@
+/**
+ * @internal
+ */
+export interface SkyIconResolved {
+  icon: string;
+  iconType: string;
+}