+++ libs/components/lists/src/lib/modules/repeater/repeater-auto-scroll-options.ts
@@ -1,0 +1,9 @@
+/**
+ * @internal
+ */
+export interface SkyRepeaterAutoScrollOptions {
+  margin: number;
+  maxSpeed: number;
+  scrollWhenOutside: boolean;
+  autoScroll(): boolean | undefined;
+}