+++ libs/components/lists/src/lib/modules/repeater/repeater.component.spec.ts
@@ -9,6 +9,8 @@ import {
 } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { SkyAppTestUtility, expect, expectAsync } from '@skyux-sdk/testing';
+import { SkyLogService } from '@skyux/core';
+import { SkyScrollableHostService } from '@skyux/core';
 import { SkyInlineFormButtonLayout } from '@skyux/inline-form';
 
 import { DragulaService, Group } from 'ng2-dragula';
@@ -20,6 +22,7 @@ import { SkyRepeaterFixturesModule } from './fixtures/repeater-fixtures.module';
 import { RepeaterInlineFormFixtureComponent } from './fixtures/repeater-inline-form.component.fixture';
 import { RepeaterWithMissingTagsFixtureComponent } from './fixtures/repeater-missing-tag.fixture';
 import { RepeaterTestComponent } from './fixtures/repeater.component.fixture';
+import { SkyRepeaterAutoScrollService } from './repeater-auto-scroll.service';
 import { SkyRepeaterComponent } from './repeater.component';
 import { SkyRepeaterService } from './repeater.service';
 
@@ -92,12 +95,48 @@ describe('Repeater item component', () => {
   function getItems(fixture: ComponentFixture<any>): HTMLElement[] {
     return fixture.nativeElement.querySelectorAll('.sky-repeater-item');
   }
+
+  function validateDeprecatedCalled(
+    deprecatedSpy: jasmine.Spy,
+    expected: boolean
+  ): void {
+    if (expected) {
+      expect(deprecatedSpy).toHaveBeenCalledOnceWith(
+        'SkyRepeaterItemComponent without `itemName`',
+        {
+          deprecationMajorVersion: 8,
+          replacementRecommendation: 'Always specify an `itemName` property.',
+        }
+      );
+    } else {
+      expect(deprecatedSpy).not.toHaveBeenCalled();
+    }
+  }
+
+  function validateRepeaterItemOrder(
+    fixture: ComponentFixture<RepeaterTestComponent>,
+    firstItemTag: string
+  ): void {
+    const cmp = fixture.componentInstance;
+    const repeaterSvc = fixture.debugElement
+      .query(By.css('sky-repeater'))
+      .injector.get(SkyRepeaterService);
+
+    const repeaterItems = cmp.repeater?.items?.toArray();
+
+    expect(repeaterItems).not.toBeUndefined();
+
+    if (repeaterItems) {
+      expect(repeaterSvc.items).toEqual(repeaterItems);
+      expect(repeaterItems[0].tag).toEqual(firstItemTag);
+      expect(repeaterSvc.items[0].tag).toEqual(firstItemTag);
+    }
+  }
   // #endregion
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [SkyRepeaterFixturesModule],
-      providers: [SkyRepeaterService],
     });
   });
 
@@ -132,6 +171,35 @@ describe('Repeater item component', () => {
 
     expect(el.querySelectorAll('sky-repeater-item').length).toBe(0);
     flushDropdownTimer();
+  }));
+
+  it('should update the repeater service items and their order items dynamically changed', fakeAsync(() => {
+    const fixture = TestBed.createComponent(RepeaterTestComponent);
+    const cmp = fixture.componentInstance;
+
+    cmp.showRepeaterWithNgFor = true;
+    detectChangesAndTick(fixture);
+
+    validateRepeaterItemOrder(fixture, 'item1');
+
+    cmp.items = [
+      {
+        id: 'item3',
+        title: 'Item 3',
+      },
+      {
+        id: 'item2',
+        title: 'Item 2',
+      },
+      {
+        id: 'item1',
+        title: 'Item 1',
+      },
+    ];
+
+    detectChangesAndTick(fixture);
+
+    validateRepeaterItemOrder(fixture, 'item3');
   }));
 
   it('should have aria-control set pointed at content', fakeAsync(() => {
@@ -170,9 +238,44 @@ describe('Repeater item component', () => {
     expect(reorderTopButtons[0].getAttribute('aria-label')).toEqual(
       'Move to top'
     );
-    expect(expandButtons[0].getAttribute('aria-label')).toEqual(
-      'Expand or collapse'
-    );
+    expect(expandButtons[0].getAttribute('aria-label')).toEqual('Expand');
+    expect(expandButtons[1].getAttribute('aria-label')).toEqual('Collapse');
+  }));
+
+  it('should warn when itemName is not defined', fakeAsync(() => {
+    const logSvc = TestBed.inject(SkyLogService);
+    const deprecatedSpy = spyOn(logSvc, 'deprecated');
+    const fixture = TestBed.createComponent(RepeaterTestComponent);
+    fixture.componentInstance.selectable = true;
+    fixture.componentInstance.reorderable = true;
+    fixture.detectChanges();
+    tick();
+    fixture.detectChanges();
+    tick();
+
+    validateDeprecatedCalled(deprecatedSpy, true);
+  }));
+
+  it('should warn when itemName is unset after initial render', fakeAsync(() => {
+    const logSvc = TestBed.inject(SkyLogService);
+    const deprecatedSpy = spyOn(logSvc, 'deprecated');
+    const fixture = TestBed.createComponent(RepeaterTestComponent);
+    fixture.componentInstance.showItemName = true; // Show item name to remove default labels
+    fixture.componentInstance.selectable = true;
+    fixture.componentInstance.reorderable = true;
+    fixture.detectChanges();
+    tick();
+    fixture.detectChanges();
+    tick();
+
+    validateDeprecatedCalled(deprecatedSpy, false);
+
+    fixture.componentInstance.showItemName = false;
+
+    fixture.detectChanges();
+    tick();
+
+    validateDeprecatedCalled(deprecatedSpy, true);
   }));
 
   it('should create aria labels when itemName is defined', fakeAsync(() => {
@@ -199,7 +302,10 @@ describe('Repeater item component', () => {
       'Move Item 1 to top'
     );
     expect(expandButtons[0].getAttribute('aria-label')).toEqual(
-      'Expand or collapse Item 1'
+      'Expand Item 1'
+    );
+    expect(expandButtons[1].getAttribute('aria-label')).toEqual(
+      'Collapse Item 2'
     );
   }));
 
@@ -1336,6 +1442,16 @@ describe('Repeater item component', () => {
     let mockDragulaService: MockDragulaService;
     let consoleSpy: jasmine.Spy;
 
+    function fireDragEvent(dragEvent: 'drag' | 'dragend', index: number): void {
+      const groupName = fixture.componentInstance.repeater?.dragulaGroupName;
+      const repeaterItem = el.querySelectorAll('sky-repeater-item')[index];
+
+      mockDragulaService[dragEvent]().next({
+        name: groupName,
+        el: repeaterItem,
+      });
+    }
+
     beforeEach(fakeAsync(() => {
       fixture = TestBed.overrideComponent(SkyRepeaterComponent, {
         add: {
@@ -1358,6 +1474,22 @@ describe('Repeater item component', () => {
         .query(By.css('sky-repeater'))
         .injector.get(DragulaService) as MockDragulaService;
     }));
+
+    function validateRepeaterItemReorderability(
+      fixture: ComponentFixture<RepeaterTestComponent>,
+      isReorderable: boolean
+    ): void {
+      const cmp = fixture.componentInstance;
+      const repeaterItems = cmp.repeater?.items?.toArray();
+
+      expect(repeaterItems).not.toBeUndefined();
+
+      if (repeaterItems) {
+        for (const item of repeaterItems) {
+          expect(item.reorderable).toBe(isReorderable);
+        }
+      }
+    }
 
     it('should not show a console warning if all item tags are defined', fakeAsync(() => {
       detectChangesAndTick(fixture);
@@ -1414,26 +1546,61 @@ describe('Repeater item component', () => {
       expect(el.querySelectorAll('sky-repeater-item')[0]).toBe(itemToTest);
     }));
 
-    it('should update css classes correctly while dragging', fakeAsync(() => {
-      const groupName = fixture.componentInstance.repeater?.dragulaGroupName;
+    it('should update css classes correctly while dragging', () => {
+      fireDragEvent('drag', 1);
+
       let repeaterItem = el.querySelectorAll('sky-repeater-item')[1];
-      mockDragulaService.drag().next({ name: groupName, el: repeaterItem });
-      fixture.detectChanges();
-      tick();
-      fixture.detectChanges();
-      repeaterItem = el.querySelectorAll('sky-repeater-item')[1];
+
       expect(
         repeaterItem.classList.contains('sky-repeater-item-dragging')
       ).toBeTruthy();
-      mockDragulaService.dragend().next({ name: groupName, el: repeaterItem });
-      fixture.detectChanges();
-      tick();
-      fixture.detectChanges();
+
+      fireDragEvent('dragend', 1);
+
       repeaterItem = el.querySelectorAll('sky-repeater-item')[1];
+
       expect(
         repeaterItem.classList.contains('sky-repeater-item-dragging')
       ).toBeFalsy();
-    }));
+    });
+
+    it('should auto-scroll while dragging', () => {
+      const repeaterInjector = fixture.debugElement.query(
+        By.css('sky-repeater')
+      ).injector;
+
+      const autoScrollSvc = repeaterInjector.get(SkyRepeaterAutoScrollService);
+      const scrollableHostSvc = repeaterInjector.get(SkyScrollableHostService);
+
+      const scrollableHostEl = scrollableHostSvc.getScrollableHost(
+        fixture.elementRef
+      );
+
+      const mockAutoScroller = jasmine.createSpyObj('autoScroller', [
+        'destroy',
+      ]);
+
+      const autoScrollSpy = spyOn(autoScrollSvc, 'autoScroll').and.returnValue(
+        mockAutoScroller
+      );
+
+      fireDragEvent('drag', 1);
+
+      expect(autoScrollSpy).toHaveBeenCalledOnceWith([scrollableHostEl], {
+        margin: 20,
+        maxSpeed: 10,
+        scrollWhenOutside: true,
+        autoScroll: jasmine.any(Function),
+      });
+
+      // The auto-scroll function should always return `true` since the auto-scroller
+      // is destroyed when dragging ends.
+      expect(autoScrollSpy.calls.argsFor(0)[1].autoScroll?.()).toBeTrue();
+
+      fireDragEvent('dragend', 1);
+
+      expect(mockAutoScroller.destroy).toHaveBeenCalledOnceWith();
+    });
 
     it('should move an item up via keyboard controls using "Space" to activate', fakeAsync(() => {
       const items = el.querySelectorAll('sky-repeater-item');
@@ -1712,6 +1879,49 @@ describe('Repeater item component', () => {
       await fixture.whenStable();
       await expectAsync(fixture.nativeElement).toBeAccessible();
     });
+
+    it('should set all items to reorderable when the repeater items change', fakeAsync(() => {
+      cmp.showRepeaterWithNgFor = true;
+      detectChangesAndTick(fixture);
+
+      validateRepeaterItemReorderability(fixture, true);
+
+      cmp.items = [
+        {
+          id: 'item4',
+          title: 'Item 4',
+        },
+        {
+          id: 'item5',
+          title: 'Item 5',
+        },
+      ];
+      detectChangesAndTick(fixture);
+
+      validateRepeaterItemReorderability(fixture, true);
+    }));
+
+    it('should show a console warning when the items change and not all item tags are defined', fakeAsync(() => {
+      cmp.showRepeaterWithNgFor = true;
+      detectChangesAndTick(fixture);
+
+      expect(consoleSpy).not.toHaveBeenCalled();
+
+      cmp.items = [
+        {
+          id: 'item4',
+          title: 'Item 4',
+        },
+        {
+          title: 'Item 5',
+        },
+      ];
+      detectChangesAndTick(fixture);
+
+      expect(consoleSpy).toHaveBeenCalledWith(
+        'Please supply tag properties for each repeater item when reordering functionality is enabled.'
+      );
+    }));
   });
 
   describe('aria roles', () => {