+++ libs/components/lists/src/lib/modules/repeater/repeater.component.ts
@@ -1,5 +1,6 @@
 import {
   AfterContentInit,
+  AfterViewChecked,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
@@ -14,13 +15,18 @@ import {
   QueryList,
   Renderer2,
   SimpleChanges,
+  inject,
 } from '@angular/core';
+import { SkyLogService } from '@skyux/core';
+import { SkyScrollableHostService } from '@skyux/core';
 
 import { DragulaService } from 'ng2-dragula';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 
 import { SkyRepeaterAdapterService } from './repeater-adapter.service';
+import { SkyRepeaterAutoScrollService } from './repeater-auto-scroll.service';
+import { SkyRepeaterAutoScroller } from './repeater-auto-scroller';
 import { SkyRepeaterExpandModeType } from './repeater-expand-mode-type';
 import { SkyRepeaterItemRolesType } from './repeater-item-roles.type';
 import { SkyRepeaterItemComponent } from './repeater-item.component';
@@ -34,11 +40,15 @@ import { SkyRepeaterService } from './repeater.service';
   selector: 'sky-repeater',
   styleUrls: ['./repeater.component.scss'],
   templateUrl: './repeater.component.html',
-  providers: [SkyRepeaterService, SkyRepeaterAdapterService],
+  providers: [
+    SkyRepeaterService,
+    SkyRepeaterAdapterService,
+    SkyRepeaterAutoScrollService,
+  ],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class SkyRepeaterComponent
-  implements AfterContentInit, OnChanges, OnDestroy, OnInit
+  implements AfterContentInit, AfterViewChecked, OnChanges, OnDestroy, OnInit
 {
   /**
    * The index of the repeater item to visually highlight as active.
@@ -107,6 +117,7 @@ export class SkyRepeaterComponent
    * This event emits an ordered array of the `tag` properties that the consumer provides for each repeater item.
    */
   @Output()
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   public orderChange = new EventEmitter<any[]>();
 
   @ContentChildren(SkyRepeaterItemComponent)
@@ -116,29 +127,21 @@ export class SkyRepeaterComponent
 
   public role: SkyRepeaterRoleType | undefined;
 
-  #adapterService: SkyRepeaterAdapterService;
-  #changeDetector: ChangeDetectorRef;
-  #dragulaService: DragulaService;
-  #elementRef: ElementRef;
-  #renderer: Renderer2;
-  #repeaterService: SkyRepeaterService;
+  #autoScroller: SkyRepeaterAutoScroller | undefined;
   #ngUnsubscribe = new Subject<void>();
+  #itemNameWarned = false;
 
-  constructor(
-    changeDetector: ChangeDetectorRef,
-    repeaterService: SkyRepeaterService,
-    adapterService: SkyRepeaterAdapterService,
-    dragulaService: DragulaService,
-    elementRef: ElementRef,
-    renderer: Renderer2
-  ) {
-    this.#changeDetector = changeDetector;
-    this.#repeaterService = repeaterService;
-    this.#adapterService = adapterService;
-    this.#dragulaService = dragulaService;
-    this.#elementRef = elementRef;
-    this.#renderer = renderer;
+  #adapterService = inject(SkyRepeaterAdapterService);
+  #changeDetector = inject(ChangeDetectorRef);
+  #dragulaService = inject(DragulaService);
+  #elementRef = inject(ElementRef);
+  #renderer = inject(Renderer2);
+  #repeaterService = inject(SkyRepeaterService);
+  #autoScrollSvc = inject(SkyRepeaterAutoScrollService);
+  #scrollableHostSvc = inject(SkyScrollableHostService);
+  #logSvc = inject(SkyLogService);
 
+  constructor() {
     this.dragulaGroupName = `sky-repeater-dragula-${
       this.#repeaterService.repeaterGroupId
     }`;
@@ -188,20 +191,19 @@ export class SkyRepeaterComponent
         this.#repeaterService.activateItemByIndex(this.activeIndex);
       }
 
-      if (this.reorderable && !this.#everyItemHasTag()) {
-        console.warn(
-          'Please supply tag properties for each repeater item when reordering functionality is enabled.'
-        );
-      }
+      this.#validateTags();
     });
 
     // HACK: Not updating for expand mode in a timeout causes an error.
     // https://github.com/angular/angular/issues/6005
     this.items?.changes.pipe(takeUntil(this.#ngUnsubscribe)).subscribe(() => {
       setTimeout(() => {
-        if (this.items?.last) {
+        if (this.items?.length) {
           this.#updateForExpandMode(this.items.last);
-          this.items.last.reorderable = this.reorderable;
+
+          this.#updateReorderability();
+
+          this.#repeaterService.items = this.items.toArray();
         }
 
         if (this.activeIndex !== undefined) {
@@ -209,6 +211,8 @@ export class SkyRepeaterComponent
         }
 
         this.#updateRole();
+
+        this.#validateTags();
       });
     });
 
@@ -221,6 +225,17 @@ export class SkyRepeaterComponent
 
       this.#updateRole();
     }, 0);
+  }
+
+  public ngAfterViewChecked(): void {
+    if (!this.#itemNameWarned && this.items?.some((item) => !item.itemName)) {
+      this.#logSvc?.deprecated('SkyRepeaterItemComponent without `itemName`', {
+        deprecationMajorVersion: 8,
+        replacementRecommendation: 'Always specify an `itemName` property.',
+      });
+
+      this.#itemNameWarned = true;
+    }
   }
 
   public ngOnChanges(changes: SimpleChanges): void {
@@ -236,7 +251,7 @@ export class SkyRepeaterComponent
 
     if (changes.reorderable) {
       if (this.items) {
-        this.items.forEach((item) => (item.reorderable = this.reorderable));
+        this.#updateReorderability();
       }
       this.#updateRole();
 
@@ -293,9 +308,11 @@ export class SkyRepeaterComponent
     /* istanbul ignore else */
     if (!this.#dragulaService.find(this.dragulaGroupName)) {
       this.#dragulaService.createGroup(this.dragulaGroupName, {
-        moves: (el, container, handle) => {
+        moves: (el, _container, handle) => {
           const target = el?.querySelector('.sky-repeater-item-grab-handle');
-          return !!this.reorderable && !!target && target.contains(handle!);
+          return !!(
+            this.reorderable && target?.contains(handle as Element | null)
+          );
         },
       });
     }
@@ -308,6 +325,18 @@ export class SkyRepeaterComponent
       .subscribe((args) => {
         /* istanbul ignore else */
         if (args.name === this.dragulaGroupName) {
+          this.#destroyAutoScroll();
+
+          this.#autoScroller = this.#autoScrollSvc.autoScroll(
+            [this.#scrollableHostSvc.getScrollableHost(this.#elementRef)],
+            {
+              margin: 20,
+              maxSpeed: 10,
+              scrollWhenOutside: true,
+              autoScroll: () => true,
+            }
+          );
+
           this.#renderer.addClass(args.el, 'sky-repeater-item-dragging');
           draggedItemIndex = this.#adapterService.getRepeaterItemIndex(
             args.el as HTMLElement
@@ -321,6 +350,8 @@ export class SkyRepeaterComponent
       .subscribe((args) => {
         /* istanbul ignore else */
         if (args.name === this.dragulaGroupName) {
+          this.#destroyAutoScroll();
+
           this.#renderer.removeClass(args.el, 'sky-repeater-item-dragging');
           const newItemIndex = this.#adapterService.getRepeaterItemIndex(
             args.el as HTMLElement
@@ -343,6 +374,15 @@ export class SkyRepeaterComponent
     /* istanbul ignore else */
     if (this.#dragulaService.find(this.dragulaGroupName)) {
       this.#dragulaService.destroy(this.dragulaGroupName);
+    }
+
+    this.#destroyAutoScroll();
+  }
+
+  #destroyAutoScroll(): void {
+    if (this.#autoScroller) {
+      this.#autoScroller.destroy();
+      this.#autoScroller = undefined;
     }
   }
 
@@ -420,6 +460,22 @@ export class SkyRepeaterComponent
       });
       this.role = `${autoRole}`;
       this.#changeDetector.markForCheck();
+    }
+  }
+
+  #updateReorderability(): void {
+    if (this.items) {
+      for (const item of this.items) {
+        item.reorderable = this.reorderable;
+      }
+    }
+  }
+
+  #validateTags(): void {
+    if (this.reorderable && !this.#everyItemHasTag()) {
+      console.warn(
+        'Please supply tag properties for each repeater item when reordering functionality is enabled.'
+      );
     }
   }
 }