+++ libs/sdk/testing/src/lib/query-predicates/sky-by.ts
@@ -2,7 +2,7 @@ import { DebugElement, Predicate } from '@angular/core';
 import { By } from '@angular/platform-browser';
 
 export class SkyBy {
-  static dataSkyId(skyId: string): Predicate<DebugElement> {
+  public static dataSkyId(skyId: string): Predicate<DebugElement> {
     return By.css(`[data-sky-id="${skyId}"]`);
   }
 }